+++ test/specs/express.spec.js
@@ -9,8 +9,9 @@ const SOME_IP = '1.3.3.7';
 const COOKIE_DEVICE_ID = 'ag_d';
 const COOKIE_SESSION_KEY = 'ag_s';
 
-const buildRequest = (cookies = {}) => ({
+const buildRequest = (cookies = {}, query = {}) => ({
   cookies,
+  query,
   ip: SOME_IP
 });
 
@@ -21,7 +22,8 @@ const buildResponse = () => ({
 
 describe('The Express middleware', () => {
   const mw = appgrid.middleware.express({
-    appKey: '56ea6a370db1bf032c9df5cb'
+    appKey: '56ea6a370db1bf032c9df5cb',
+    gid: 'myGID',
   });
 
   it('exposes a function', () => {
@@ -44,8 +46,24 @@ describe('The Express middleware', () => {
     const { appgridClient } = response.locals;
     appgridClient.should.be.an('object');
     appgridClient.props.config.appKey.should.equal('56ea6a370db1bf032c9df5cb');
+    appgridClient.props.config.gid.should.equal('myGID');
     appgridClient.props.config.deviceId.should.equal(cookies[COOKIE_DEVICE_ID]);
     appgridClient.props.config.sessionKey.should.equal(cookies[COOKIE_SESSION_KEY]);
+  });
+
+  it('should use the gid from the request if any was found there (even if a gid was set in the middleware config)', () => {
+    const response = buildResponse();
+    const cookies = {
+      [COOKIE_DEVICE_ID]: 'gregTestingSDK',
+      [COOKIE_SESSION_KEY]: 'gregSessionKey'
+    };
+    const REQUEST_GID = 'a_cool_gid';
+    const query = {
+      gid: REQUEST_GID
+    };
+    mw(buildRequest(cookies, query), response, NOOP);
+    const { appgridClient } = response.locals;
+    appgridClient.props.config.gid.should.equal(REQUEST_GID);
   });
 
   describe('when no deviceId or sessionKey is found in the request cookies', () => {